Fix integer overflow in RangeDataset.
PiperOrigin-RevId: 491648273

+++ tensorflow/core/kernels/data/range_dataset_op.cc
@@ -80,7 +80,7 @@ class RangeCounter {
       return -1;
     }
     *end_of_counter = false;
-    int result = next_;
+    int64_t result = next_;
     next_ += step_;
     return result;
   }
